+++ processor/src/test/fixtures/bad_input/com/example/package_access/other/BaseOnActivateModel.java
@@ -3,7 +3,6 @@ package com.example.package_access.other;
 import arez.annotations.Memoize;
 import arez.annotations.OnActivate;
 import arez.annotations.OnDeactivate;
-import arez.annotations.OnStale;
 
 public abstract class BaseOnActivateModel
 {
@@ -20,11 +19,6 @@ public abstract class BaseOnActivateModel
 
   @OnDeactivate
   public final void onTimeDeactivate()
-  {
-  }
-
-  @OnStale
-  public final void onTimeStale()
   {
   }
 }