+++ processor/src/test/fixtures/input/com/example/collections/MemoizeCollectionWithHooksModel.java
@@ -4,7 +4,6 @@ import arez.annotations.ArezComponent;
 import arez.annotations.Memoize;
 import arez.annotations.OnActivate;
 import arez.annotations.OnDeactivate;
-import arez.annotations.OnStale;
 import java.util.Collection;
 import java.util.Collections;
 
@@ -24,11 +23,6 @@ abstract class MemoizeCollectionWithHooksModel
 
   @OnDeactivate
   final void onTimeDeactivate()
-  {
-  }
-
-  @OnStale
-  final void onTimeStale()
   {
   }
 }