+++ processor/src/test/fixtures/input/com/example/inheritance/other/BaseCompleteModel.java
@@ -23,7 +23,6 @@ import arez.annotations.ObserverRef;
 import arez.annotations.OnActivate;
 import arez.annotations.OnDeactivate;
 import arez.annotations.OnDepsChange;
-import arez.annotations.OnStale;
 import arez.annotations.PostConstruct;
 import arez.annotations.PostDispose;
 import arez.annotations.PreDispose;
@@ -96,11 +95,6 @@ public abstract class BaseCompleteModel
 
   @OnDeactivate
   protected final void onTimeDeactivate()
-  {
-  }
-
-  @OnStale
-  protected final void onTimeStale()
   {
   }