+++ processor/src/test/fixtures/input/com/example/memoize/WithHooksModel.java
@@ -4,7 +4,6 @@ import arez.annotations.ArezComponent;
 import arez.annotations.Memoize;
 import arez.annotations.OnActivate;
 import arez.annotations.OnDeactivate;
-import arez.annotations.OnStale;
 
 @ArezComponent
 abstract class WithHooksModel
@@ -22,11 +21,6 @@ abstract class WithHooksModel
 
   @OnDeactivate
   final void onTimeDeactivate()
-  {
-  }
-
-  @OnStale
-  final void onTimeStale()
   {
   }
 }